chore(structure) Begin layout
