+++ src/README.md
@@ -1,0 +1,3 @@
+Digital Bearer Certificates for the Safe network. 
+
+Docs to follow.
