+++ src/lib.rs
@@ -1,11 +1,12 @@
 use serde::{Deserialize, Serialize};
-use serde_json::Result;
+ase serde_json::Result;
 use tiny_keccak::{Hasher, Sha3};
 /// These typdefs are to simplify algorithm for now and will be removed for production.
-type Hash = [u8; 32];
-type PubKey = [u8; 32]; // tmp
-type DbcContentHash = [u8; 32];
-type DbcSpentHash = [u8; 32];
+pub(crate) type Hash = [u8; 32];
+pub(crate)type PubKey = [u8; 32]; // tmp
+pub(crate)type DbcContentHash = [u8; 32];
+pub(crate)type DbcSpentHash = [u8; 32];
+mod dbc;
 
 fn sha3_256(input: &[u8]) -> Hash {
     let mut sha3 = Sha3::v256();
@@ -15,22 +16,19 @@ fn sha3_256(input: &[u8]) -> Hash {
     output
 }
 
-/// This is the content of a DBC, it is unique as the parent hash is included
+/// This is the content of a DBC, it is unique as the parent hash is included 
 #[derive(Serialize, Deserialize)]
 struct DbcContent {
     parent: DbcContentHash, // Hash of parent DbcContent. Also used as a nonce
     owner: PubKey,          // Will be blinded by using Hash(Hash(PubKey + self.parent))
     amount: u64,
 }
-// TODO, we may wish to add a nonce such that
-// All output DBCs hash to have no common leading bits
-// i.e. log base2 (num_outputs) == num leading bits that must be differnt
-// This means a DBCs parent has to have been in a differnt section
+
 impl DbcContent {
     // Create a new DbcContent for signing, blind the owner from the mint
     pub fn new(parent: DbcContentHash, owner: PubKey, amount: u64) -> Self {
         let mut owner = owner;
-        for _ in 0..amount % 1000 {
+        for _ in 0..amount {
             owner = sha3_256(&owner); // owner not visible to mint.
         }
         DbcContent {
@@ -45,28 +43,7 @@ impl DbcContent {
     }
 }
 
-struct Dbc {
-    content: DbcContent,
-    parent_spent: DbcSpent,
-    mint_key: PubKey,
-    mint_sig: Signature, // (self.parent_spent),
-}
-impl Dbc {
-    // Take an input DBC and return output dbcContent(s)
-    // To create a valid DBC the Mint must sign this after doing some checks.
-    // verify that the dbc contents was part of the parent spend transacition.
-    //     hash(dbc.content) ∈ dbc.parent_spent.outputs
-    //     dbc.content.parent == dbc.parent_spent.input
 
-    // If we don’t recognize the dbc.section_key, fetch a proof chain from the section
-    // responsible for XorName(hash(dbc.parent_spent.input)) and validated the proof chain
-    //  as well as that dbc.section_key is present in the chain.
-    // Cache the dbc.section_key
-    // Verify the dbc.section_sig signature is a valid signature of dbc.parent_spent.
-    // dbc.section_key.verify(dbc.section_sig, hash(dbc.parent_spent))
-    // Outputs signed with BLS public key share
-    pub fn mint(input: Dbc, outputs: Vec<Dbc>) -> Result<Error, Vec<Self>> {}
-}
 /// The spent identifier of the outputs created from this input
 /// Note these are hashes and not identifiers as the Dbc is not addressable on the network.
 /// i.e. a Dbc can be stored anywhere, even offline.
